Yoinked code from Damus to enable event signing and settings sync (#5)

+++ Primal/Common/Model/Nostr.swift
@@ -9,7 +9,7 @@ import Foundation
 import GenericJSON
 
 struct NostrContent: Codable {
-    let kind: Int8
+    let kind: Int32
     let content: String
     let id: String
     let created_at: Int32
@@ -17,7 +17,7 @@ struct NostrContent: Codable {
     let sig: String
     let tags: [[String]]
     
-    init(kind: Int8, content: String, id: String, created_at: Int32, pubkey: String, sig: String, tags: [[String]]) {
+    init(kind: Int32, content: String, id: String, created_at: Int32, pubkey: String, sig: String, tags: [[String]]) {
         self.kind = kind
         self.content = content
         self.id = id
@@ -28,7 +28,7 @@ struct NostrContent: Codable {
     }
     
     init(json: JSON) {
-        self.kind = Int8(json.arrayValue?[2].objectValue?["kind"]?.doubleValue ?? -1)
+        self.kind = Int32(json.arrayValue?[2].objectValue?["kind"]?.doubleValue ?? -1)
         self.content = json.arrayValue?[2].objectValue?["content"]?.stringValue ?? ""
         self.id = json.arrayValue?[2].objectValue?["id"]?.stringValue ?? ""
         self.created_at = Int32(json.arrayValue?[2].objectValue?["created_at"]?.doubleValue ?? -1)
